WIP: Clarify that permission_rationale_post_notifications is about posting new notifications and not about accessing existing ones
I was initially quite confused when I first read it. Because there's a
seperate special permission for an app to access existing notifications
(and I had to interact with that recently), it took me a solid while to
understand that this is really just about _sending_ notifications, like
most normal apps do.

TODO: Is this actually necessary?
TODO: Maybe just squash it into that big "improvements" commit?
TODO: If other languages could use a similar fixup, how should that be
      handled? Within this commit? Independently, but still somehow
      announced? Or maybe it's better to just let others figure it out
      on their own?